Move new scripts to the helper folder
